renamed stateShuffler files to Shuffler
