+++ CMakeLists.txt
@@ -6,7 +6,7 @@
 #    By: mfiguera <mfiguera@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2020/09/23 10:17:31 by jpinyot           #+#    #+#              #
-#    Updated: 2020/10/01 15:37:47 by mfiguera         ###   ########.fr        #
+#    Updated: 2020/10/05 11:05:51 by mfiguera         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -30,7 +30,7 @@ add_executable(n-puzzle
 			srcs/stateManhattanDistance.cpp
 			srcs/closedStack.cpp
 			srcs/openStack.cpp
-			srcs/stateShuffler.cpp
+			srcs/shuffler.cpp
 			)
 
 target_include_directories(n-puzzle PRIVATE 
@@ -43,5 +43,5 @@ set(HEADER_FILES
 			includes/nodes.h
 			includes/stateManhattanDistance.h
 			includes/openStack.h
-			includes/stateShuffler.h
+			includes/shuffler.h
 			)
